% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_functions.R
\name{LR_calculation_FS}
\alias{LR_calculation_FS}
\title{Calculate likelihood ratio for Full Sibling}
\usage{
LR_calculation_FS(personA, personB, allele_fre, r)
}
\arguments{
\item{personA}{First person's genotype}

\item{personB}{Second person's genotype}

\item{allele_fre}{Allele frequency data frame}

\item{r}{Relationship coefficient}
}
\value{
A matrix with likelihood ratios
}
\description{
Calculate likelihood ratio for Full Sibling
}
