% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{LR_excl_SK}
\alias{LR_excl_SK}
\title{Calculate likelihood ratio for unrelated individuals (r=1/4)}
\usage{
LR_excl_SK(freq, n = 10000)
}
\arguments{
\item{freq}{Allele frequency data frame}

\item{n}{Number of simulations (default: 10000)}
}
\value{
A matrix with likelihood ratios
}
\description{
Calculate likelihood ratio for unrelated individuals (r=1/4)
}
