% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{LR_incl_GP}
\alias{LR_incl_GP}
\title{Calculate likelihood ratio for Grandparent-Grandchild relationship}
\usage{
LR_incl_GP(freq, n = 10000)
}
\arguments{
\item{freq}{Allele frequency data frame}

\item{n}{Number of simulations (default: 10000)}
}
\value{
A matrix with likelihood ratios
}
\description{
Calculate likelihood ratio for Grandparent-Grandchild relationship
}
