% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{LR_incl_HS}
\alias{LR_incl_HS}
\title{Calculate likelihood ratio for Half-Sibling relationship}
\usage{
LR_incl_HS(freq, n = 10000)
}
\arguments{
\item{freq}{Allele frequency data frame}

\item{n}{Number of simulations (default: 10000)}
}
\value{
A matrix with likelihood ratios
}
\description{
Calculate likelihood ratio for Half-Sibling relationship
}
