% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{produce_offspring}
\alias{produce_offspring}
\title{Create offspring from two parents}
\usage{
produce_offspring(father, mother)
}
\arguments{
\item{father}{A data frame representing father's genotype}

\item{mother}{A data frame representing mother's genotype}
}
\value{
A data frame representing offspring's genotype
}
\description{
Create offspring from two parents
}
