% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randomperson}
\alias{randomperson}
\title{Create a random person with genotypes}
\usage{
randomperson(freq)
}
\arguments{
\item{freq}{A data frame containing allele frequencies}
}
\value{
A data frame representing a person's genotype
}
\description{
Create a random person with genotypes
}
